Crate the activity that control the  product_detail_activity layout.
Map the layouts' element to the variables in activity, set buttons' method.

Use the state system to control the layout, this help reuse the layout and reduce activity, one layout can display the detail, use to adding or editing one item by revealing or hiding the element. The change state methods is responsible for changing the layout